+++ CHANGELOG.md
@@ -19,6 +19,11 @@
 - Added better integration of Adorn's Candlelit Lanterns.
 - Fixed some cases where block state IDs would not be able to sync back up.
 
+### 1.0.2
+
+- Improved support of oxidizing and waxable lanterns.
+- Improved wall lanterns on walls that do not have the post part.
+
 ## 1.1.0
 
 - Ported to Minecraft 1.21.1.
@@ -26,5 +31,7 @@
 
 ### 1.1.1
 
+- Improved support of oxidizing and waxable lanterns.
+- Improved wall lanterns on walls that do not have the post part.
 - Fixed missing recipes for the Amethyst and Redstone Lanterns.
 - Fixed missing `#auroraslanterns:wall_lanterns` block tag.